+++ app/page.tsx
@@ -8,25 +8,10 @@ import { INTERNAL_ROLES } from "@/lib/access";
 
 import { Countdown } from "./components/Countdown";
 import { FloatingMenu } from "./components/FloatingMenu";
+import HeaderSlider from "./components/HeaderSlider";
 import { getNews } from "./news/newsData";
 import { NewsItem } from "./news/newsItem";
 import styles from "./top-page.module.css";
-
-const images = [
-  "/background/background (1).jpg",
-  "/background/background (2).jpg",
-  "/background/background (3).jpg",
-  "/background/background (4).jpg",
-  "/background/background (5).jpg",
-  "/background/background (6).jpg",
-  "/background/background (7).jpg",
-  "/background/background (8).jpg",
-  "/background/background (9).jpg",
-  "/background/background (10).jpg",
-  "/background/background (11).jpg",
-  "/background/background (12).jpg",
-  "/background/background (13).jpg",
-];
 
 const latestNews = getNews()
   .sort((a, b) => (a.date < b.date ? 1 : -1))
@@ -58,11 +43,7 @@ export default function Toppage() {
     <>
       <header className={styles.header}>
         <div className={styles.bgSlider}>
-          {images.map((src, index) => (
-            <div key={src} className={styles.slide}>
-              <img src={src} loading="lazy" alt="" aria-hidden="true" />
-            </div>
-          ))}
+          <HeaderSlider />
         </div>
         <div className={styles.themeContainer}>
           <Image
