+++ app/top-page.module.css
@@ -33,87 +33,6 @@
   height: 100%;
   z-index: 0;
 }
-.slide {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  opacity: 0;
-  animation: fadeSlideshow 52s linear infinite;
-}
-.slide:nth-child(1) {
-  animation-delay: 0s;
-}
-.slide:nth-child(2) {
-  animation-delay: -4s;
-}
-.slide:nth-child(3) {
-  animation-delay: -8s;
-}
-.slide:nth-child(4) {
-  animation-delay: -12s;
-}
-.slide:nth-child(5) {
-  animation-delay: -16s;
-}
-.slide:nth-child(6) {
-  animation-delay: -20s;
-}
-.slide:nth-child(7) {
-  animation-delay: -24s;
-}
-.slide:nth-child(8) {
-  animation-delay: -28s;
-}
-.slide:nth-child(9) {
-  animation-delay: -32s;
-}
-.slide:nth-child(10) {
-  animation-delay: -36s;
-}
-.slide:nth-child(11) {
-  animation-delay: -40s;
-}
-.slide:nth-child(12) {
-  animation-delay: -44s;
-}
-.slide:nth-child(13) {
-  animation-delay: -48s;
-}
-.slide img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  object-position: center;
-}
-@media (prefers-reduced-motion: reduce) {
-  .slide {
-    animation: none;
-  }
-  .skude:first-child {
-    opacity: 1;
-  }
-}
-
-@keyframes fadeSlideshow {
-  0% {
-    opacity: 0;
-  }
-  1.92% {
-    opacity: 1;
-  }
-  7.69% {
-    opacity: 1;
-  }
-  9.61% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 0;
-  }
-}
-
 .themeContainer {
   position: relative;
   display: inline-block;
