+++ content/changelog/0284-changelog.json
@@ -1,5 +1,5 @@
 {
   "title": "ヘッダーの背景画像のダウンロード",
-  "description": "ヘッダーの背景画像を変更・ダウンロードしました。",
+  "description": "ヘッダースライダーのコンポーネントを追加し、背景画像を変更・ダウンロードしました。",
   "credits": ["@kinoto0103"]
 }
